Refactor Groovy assert/bind a/ base classes; add JavaScript variants
